add attribute driver
